Change Description : Added the code for configuring the vm for parameters like cpuAllocation & memoryAllocation for reservation & limit ( storag_ &_VM_Checks )
Unit Testing : Due to some specific value configuration for the above parameters .. VSphere triggers the error saying - 'spec.memoryAllocation.reservation parameter is not correct'.
Not sure what  value to set & how it is relevant to guest allocated memory  & other parameters [ Need to explore more for this].
But apart from knowing the values functionality is in place.